User system and my page

+++ member/views.py
@@ -7,13 +7,29 @@ from django.utils import timezone
 from django.http import HttpResponse, request
 from django.core.paginator import Paginator
 from django.contrib.auth import login as auth_login
+from django.contrib.auth.forms import AuthenticationForm
 
 #bottle - django 
 from django.contrib.auth.views import (
     LoginView, logout_then_login
 )
 
-login = LoginView.as_view(template_name="member/login_form.html")
+# login = LoginView.as_view(template_name="member/login_form.html")
+
+def login(request) :
+    if request.method == 'POST' :
+        form = AuthenticationForm(request, data = request.POST)
+        if form.is_valid() :
+            auth_login(request, form.get_user())
+            return redirect('codagram:index')
+    else :
+        form = AuthenticationForm()
+
+    context = {
+        'form' : form,
+    }
+    return render(request, 'member/login_form.html', context)
+
 
 def signup(request):
     if request.method == 'POST':
@@ -33,14 +49,14 @@ def logout(request) :
     return logout_then_login(request)
 
 
-def my_page(request) :
+def my_paper(request) :
     # 내가 쓴 글 리스트업 작업
     wri_page = request.GET.get('wri_page', 1)
 
-    #최신순 정렬이 필요해보입니다.
-
-    id = request.session['user_name']
+    id = request.user.get_username()
     write = Write.objects.filter(author = id)
+
+    write.order_by('-update_time', '-make_time')
 
     write_p = Paginator(write, 5)
     wri_info = write_p.page(wri_page)
@@ -54,6 +70,8 @@ def my_page(request) :
     com_page = request.GET.get('com_page', 1)
 
     comment = Comment.objects.filter(author = id)
+
+    comment.order_by('-make_time','-update_time')
 
     com_p = Paginator(comment, 5)
     com_info = com_p.page(com_page)
@@ -70,17 +88,39 @@ def my_page(request) :
         'com_range' : range(com_start_page, com_end_page)
     }
 
+    return context
+
+def my_page(request) :
+    # mypage에 들어갈 데이터 조회
+    context = my_paper(request)
+
     return render(request, 'member/mypage.html', context)
 
 def delete(request) :
+    # mypage에서 댓글 삭제나 글 삭제 기능
+
     text_type = request.GET.get('text_type')
     id = request.GET.get('id')
     
-    if text_type :
+    if text_type == 1 :
         data = Comment.objects.filter(id=id)
+
     else :
         data = Write.objects.filter(id=id)
     
-    data.delete()
+    if data.author == request.user.get_username() :
+        data.delete()
 
     return redirect('member:mypage')
+
+# def alarm(request) :
+#     # 사용자의 알람을 조회하는 기능
+
+#     user = User.objects.filter(username=request.user.get_username())
+#     alarm = Alarm.objects.filter(user = user)
+    
+#     context = {
+#         'alarm' : alarm
+#     }
+
+#     return alarm
